api/requirements.txt
fastapi==0.110.0
uvicorn==0.28.0
python-multipart==0.0.9
Pillow==10.2.0
torch==2.2.1+cpu 
torchvision==0.17.1+cpu
numpy==1.26.4
opencv-python-headless==4.9.0.80
--extra-index-url https://download.pytorch.org/whl/cpu